refactor: user 라우터 uri 복수형으로 수정
